requirements.txt
fastapi
uvicorn
python-docx
chromadb
groq
sentence-transformers
python-multipart
python-dotenv
